backend-crt/src/requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cssselect==1.2.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.4
google-search-results==2.4.2
idna==3.4
jieba3k==0.35.1
joblib==1.3.2
lxml==4.9.3
newspaper3k==0.2.8
nltk==3.8.1
Pillow==10.0.1
psycopg2==2.9.7
pydantic==1.10.13
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
tinysegmenter==0.3
tldextract==3.6.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.5
